Refactor containerservice to utils
