+++ st2reactor/tests/test_container_utils.py
@@ -2,10 +2,9 @@ import datetime
 import mock
 import tests
 import unittest2
-import st2reactor.container.containerservice
 from st2common.persistence.reactor import Trigger, TriggerInstance
 from st2common.models.db.reactor import TriggerDB, TriggerInstanceDB
-from st2reactor.container import containerservice
+from st2reactor.container import utils
 
 MOCK_TRIGGER = TriggerDB()
 MOCK_TRIGGER.id = 'trigger-test.id'
@@ -27,9 +26,9 @@ class ContainerServiceTest(unittest2.TestCase):
         return_value=[MOCK_TRIGGER]))
     @mock.patch.object(TriggerInstance, 'add_or_update', mock.MagicMock(
         return_value=MOCK_TRIGGER_INSTANCE))
-    @mock.patch('st2reactor.container.containerservice.DISPATCH_HANDLER')
+    @mock.patch('st2reactor.container.utils.DISPATCH_HANDLER')
     def test_validate_dispatch(self, mock_dispatch_handler):
-        containerservice.dispatch_trigger(MOCK_TRIGGER_INSTANCE)
+        utils.dispatch_trigger(MOCK_TRIGGER_INSTANCE)
         mock_dispatch_handler.assert_called_once_with([MOCK_TRIGGER_INSTANCE])
 
     @mock.patch.object(Trigger, 'query', mock.MagicMock(
@@ -37,5 +36,5 @@ class ContainerServiceTest(unittest2.TestCase):
     @mock.patch.object(Trigger, 'add_or_update')
     def test_add_trigger(self, mock_add_handler):
         mock_add_handler.return_value = MOCK_TRIGGER
-        containerservice.add_trigger_type(MOCK_TRIGGER)
+        utils.add_trigger_type(MOCK_TRIGGER)
         self.assertTrue(mock_add_handler.called, 'trigger not added.')
